requirements.txt
numpy>=1.24.0
scipy>=1.10.0
mujoco>=3.1.0
pyserial>=3.5
bleak>=0.21.0
matplotlib>=3.7.0
websockets>=11.0
transforms3d>=0.4.1
PyOpenGL>=3.1.0
glfw>=2.5.0
